refactor: need access to the id of all models with a 1:N or 1:1 relationship
add IdentifiedModel
change id of LearnModel from Int to Long

+++ ui/src/main/java/bx/app/ui/ScreenManager.kt
@@ -70,7 +70,7 @@ class ScreenManager(private var context: Context, private val topBarViewModel: T
 
     @Composable
     fun DeckLearn(
-        onClickLearn: (id: Int) -> Unit = {},
+        onClickLearn: (id: Long) -> Unit = {},
     ) {
         DeckLearnScreen(context, topBarViewModel, onClickLearn)
     }
